homework1/requirements.txt
Pillow
matplotlib
numpy
torch>=2.1.2+cu118
torchvision>=0.15.0+cu117
